✨ Feat: 공통 옵션 분리 및 prefetch 쿼리 적용
