+++ src/shared/lib/api.ts
@@ -1,3 +1,5 @@
+import { PAGE_SIZE } from '@/shared/constants/pageSize';
+import { QUERY_KEYS } from '@/shared/constants/queryKey';
 import { HackerNewsItem, StoryTab } from '@/shared/types/hackerNews';
 
 const BASE_URL = 'https://hacker-news.firebaseio.com/v0';
@@ -17,3 +19,16 @@ export const fetchStory = async (id: number): Promise<HackerNewsItem> => {
   }
   return res.json();
 };
+
+export const storiesQueryOptions = (tab: StoryTab) => ({
+  queryKey: [QUERY_KEYS.stories, tab],
+  queryFn: async ({ pageParam }: { pageParam: number }) => {
+    const allIds = await fetchStoryIds(tab);
+    const pageIds = allIds.slice(
+      pageParam * PAGE_SIZE,
+      (pageParam + 1) * PAGE_SIZE,
+    );
+    return Promise.all(pageIds.map(fetchStory));
+  },
+  initialPageParam: 0,
+});
